refactor: twitter service
