+++ app/services/twitter/oauth2.py
@@ -1,70 +1,11 @@
 """Twitter OAuth2 authentication module."""
 
-from urllib.parse import urlencode
-
 from fastapi import APIRouter, Depends
 from pydantic import BaseModel
-from requests.auth import HTTPBasicAuth
-from requests_oauthlib import OAuth2Session
 
 from app.auth import verify_admin_jwt
+from intentkit.clients.twitter import OAuth2UserHandler
 from intentkit.config.config import config
-
-
-# this class is forked from:
-# https://github.com/tweepy/tweepy/blob/main/tweepy/auth.py
-# it is not maintained by the original author, bug need to be fixed
-class OAuth2UserHandler(OAuth2Session):
-    """OAuth 2.0 Authorization Code Flow with PKCE (User Context)
-    authentication handler
-    """
-
-    def __init__(self, *, client_id, redirect_uri, scope, client_secret=None):
-        super().__init__(client_id, redirect_uri=redirect_uri, scope=scope)
-        if client_secret is not None:
-            self.auth = HTTPBasicAuth(client_id, client_secret)
-        else:
-            self.auth = None
-        self.code_challenge = self._client.create_code_challenge(
-            self._client.create_code_verifier(128), "S256"
-        )
-
-    def get_authorization_url(self, agent_id: str, redirect_uri: str):
-        """Get the authorization URL to redirect the user to
-
-        Args:
-            agent_id: ID of the agent to authenticate
-            redirect_uri: URI to redirect to after authorization
-        """
-        state_params = {"agent_id": agent_id, "redirect_uri": redirect_uri}
-        authorization_url, _ = self.authorization_url(
-            "https://x.com/i/oauth2/authorize",
-            state=urlencode(state_params),
-            code_challenge=self.code_challenge,
-            code_challenge_method="S256",
-        )
-        return authorization_url
-
-    def get_token(self, authorization_response):
-        """After user has authorized the app, fetch access token with
-        authorization response URL
-        """
-        return super().fetch_token(
-            "https://api.x.com/2/oauth2/token",
-            authorization_response=authorization_response,
-            auth=self.auth,
-            include_client_id=True,
-            code_verifier=self._client.code_verifier,
-        )
-
-    def refresh(self, refresh_token: str):
-        """Refresh token"""
-        return super().refresh_token(
-            "https://api.x.com/2/oauth2/token",
-            refresh_token=refresh_token,
-            include_client_id=True,
-        )
-
 
 # Initialize Twitter OAuth2 client
 oauth2_user_handler = OAuth2UserHandler(
